tokener ok, ft_realloc ok, Norm ok
