+++ src/ft_realloc.c
@@ -6,7 +6,7 @@
 /*   By: mgagnon <mgagnon@student.42quebec.com      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/04 10:47:04 by mgagnon           #+#    #+#             */
-/*   Updated: 2023/04/05 20:41:57 by mgagnon          ###   ########.fr       */
+/*   Updated: 2023/04/06 09:45:10 by mgagnon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ void	*ft_realloc(void *ptr, size_t size)
 	new_ptr = malloc(size);
 	if (new_ptr == NULL)
 		return (NULL);
-	ft_memcpy(new_ptr, ptr, size);
+	ft_memcpy(new_ptr, ptr, sizeof(ptr));
 	free(ptr);
 	return (new_ptr);
 }
